Create codeql.yml
The iModel.js org has access to the Code Scanning beta offered by GitHub.  Setup a basic scan that will run on every PR and on a weekly basis.

This is very similar, if not the same, as what will be found on LGTM, https://lgtm.com/projects/g/imodeljs/imodeljs-samples/?mode=list. 

Learn more about how to config it and monitor the results [here](https://docs.github.com/en/github/finding-security-vulnerabilities-and-errors-in-your-code).